pyOptimizer/requirements.txt
Flask
python-dotenv
requests
kubernetes